refactor(lib): replace Windows types with standard C++ types and improve time precision
- replaced Windows-specific types (`ULONGLONG`, `UINT`, `BYTE`, `WCHAR`) with standard C++ types (`uint64_t`, `uint32_t`, `uint8_t`, `wchar_t`).
- refactored `ProcessEvent` to pass `Object^ payload` directly, removing `WParam`/`LParam` and unsafe reinterpret casts.
- migrated `elapsed_time`, `length`, and time-related APIs from `float` to `double` for better precision.
- added gfortran to CI workflow, disabled fftw3 dependency due to licensing problem.
- updated `README.md` description and added FAudio license.

+++ WpfMusicPlayer/ViewModels/LyricsViewModel.cs
@@ -48,8 +48,8 @@ public partial class LyricsViewModel(
     public event Action<string, int>? UpdateCurrentLyricRequested;
     public event Action<int>? UpdateCurrentLyricOffsetRequested;
 
-    private float _currentLyricDuration;
-    private float _lastPlaybackTimeSec;
+    private double _currentLyricDuration;
+    private double _lastPlaybackTimeSec;
     private int _currentLyricOffsetMs;
     private string? _currentIntermediateLyricJson;
     private string _suggestedWplrcFileName = "lyrics";
@@ -119,7 +119,7 @@ public partial class LyricsViewModel(
         ExportWplrcCommand.NotifyCanExecuteChanged();
     }
 
-    public void UpdateLyricProgress(float time)
+    public void UpdateLyricProgress(double time)
     {
         _lastPlaybackTimeSec = time;
         if (_lyricStates.Count == 0) return;
@@ -166,7 +166,7 @@ public partial class LyricsViewModel(
         string? filePath,
         string? suppliedLyric,
         string? songTitle,
-        float songDuration,
+        double songDuration,
         int offsetMs,
         SuppliedLyricSource suppliedLyricSource = SuppliedLyricSource.Embedded)
     {
@@ -247,7 +247,7 @@ public partial class LyricsViewModel(
         OnPropertyChanged(nameof(CurrentLyric));
     }
 
-    private double CalculateLinearLyricProgress(int index, float currentTimeSec)
+    private double CalculateLinearLyricProgress(int index, double currentTimeSec)
     {
         if (index < 0 || index >= Lyrics.Count) return 0;
